bloc(cubit) added and little updates

+++ lib/pages/user_list_page.dart
@@ -5,6 +5,8 @@ import 'package:flutter_mongo_example/pages/user_create_page.dart';
 import 'package:flutter_mongo_example/pages/user_update_page.dart';
 import 'package:flutter_mongo_example/models/user_model.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_mongo_example/widgets/custom_alert_dialog.dart';
+import 'package:flutter_mongo_example/widgets/custom_text_field.dart';
 
 class UserListPage extends StatefulWidget {
   const UserListPage({Key? key}) : super(key: key);
@@ -29,19 +31,18 @@ class _UserListPageState extends State<UserListPage> {
         tooltip: "Add",
         child: const Icon(Icons.add),
         onPressed: () {
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) {
-                return UserCreatePage(
-                  onSuccess: () {
-                    cubit.getData();
-                    Navigator.pop(context);
-                  },
-                );
-              },
-            ),
-          );
+          showDialog(
+              context: context,
+              builder: (context) => AlertDialog(
+                    shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(20)),
+                    content: UserCreatePage(
+                      onSuccess: () {
+                        cubit.getData();
+                        Navigator.pop(context);
+                      },
+                    ),
+                  ));
         },
       ),
       appBar: AppBar(
@@ -55,11 +56,21 @@ class _UserListPageState extends State<UserListPage> {
           child: BlocBuilder<UserListCubit, List<UserModel>>(
             bloc: cubit,
             builder: (context, state) {
-              return ListView.builder(
-                itemCount: state.length,
-                itemBuilder: (context, index) {
-                  return getCard(state[index], context, cubit);
-                },
+              return Column(
+                children: [
+                  CustomTextField(
+                    hintText: "Search",
+                    onChanged: cubit.getDataByFilter,
+                  ),
+                  Expanded(
+                    child: ListView.builder(
+                      itemCount: state.length,
+                      itemBuilder: (context, index) {
+                        return getCard(state[index], context, cubit);
+                      },
+                    ),
+                  ),
+                ],
               );
             },
           ),
@@ -73,33 +84,43 @@ Card getCard(UserModel user, BuildContext context, UserListCubit cubit) {
   return Card(
     child: ListTile(
       title: Text(user.name),
-      subtitle: Text(user.password),
+      subtitle: Text(user.phone),
       trailing: Row(
         mainAxisSize: MainAxisSize.min,
         children: [
           IconButton(
               tooltip: "Edit",
               onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                      builder: (context) {
-                        return UserUpdatePage(
-                          onSuccess: () {
-                            cubit.getData();
-                            Navigator.pop(context);
-                          },
-                        );
-                      },
-                      settings: RouteSettings(arguments: user)),
-                );
+                showDialog(
+                    context: context,
+                    builder: (context) => AlertDialog(
+                          shape: RoundedRectangleBorder(
+                              borderRadius: BorderRadius.circular(20)),
+                          content: UserUpdatePage(
+                            user: user,
+                            onSuccess: () {
+                              cubit.getData();
+                              Navigator.pop(context);
+                            },
+                          ),
+                        ));
               },
               icon: const Icon(Icons.edit)),
           IconButton(
               tooltip: "Delete",
               onPressed: () async {
-                await MongoDatabase.delete(user);
-                cubit.getData();
+                showDialog(
+                  context: context,
+                  builder: (context) {
+                    return CustomAlertDialog(
+                      user: user,
+                      onSuccess: () async {
+                        await MongoDatabase.delete(user);
+                        cubit.getData();
+                      },
+                    );
+                  },
+                );
               },
               icon: const Icon(Icons.delete)),
         ],
